Added mode to shared preferences & Animated
